[Xcode] Use scheme scoping to build WebKit separate from internal projects https://bugs.webkit.org/show_bug.cgi?id=292962 rdar://149564758
Reviewed by NOBODY (OOPS!).

Prepare for adoption of scheme scoping [1] in "All WebKit Tools" and the
family of "Everything up to WebKit" schemes. This setting causes Xcode
to filter the build graph to only include targets that are listed in the
active scheme [2]. For WebKit, which can build internally as part of
Safari, this change makes it so that:

- Building WebKit never pulls in part of the Safari stack, adding to
  overall build time; and
- Whether a build of WebKit uses Safari frameworks from the SDK or from
  ToT no longer depends on the active workspace.

Scheme scoping requires that all our targets are either part of the
scheme, or in an aggregate target that's part of the scheme. This means
that some targets which were only added to the build graph transitively
or implicitly must be explicit declared, usually in a project's "All"
aggregate target.

Some projects did not have an "All" aggregate target, so add it there.

[1]: We cannot use scheme scoping in public builds yet because of a bug
related to DEPENDENCY_SCOPE_INCLUDES_DIRECT_DEPENDENCIES with public
Xcode (rdar://136703410) until Xcode 16.3. Because we support older
versions, these project changes will only benefit internal-only build
configurations for now. Schemes in WebKit.xcworkspace continue to be
unscoped, and this change should not impact their build order.

* Configurations/CommonBase.xcconfig:
[2]: By default, scheme scoping does not consider dependencies of
aggregate targets to be part of the scope. This does not work for
WebKit, because we use aggregate targets to make it possible to
introduce new targets to the build without needing to update every
scheme across all workspaces. Customize this behavior by setting
DEPENDENCY_SCOPE_INCLUDES_DIRECT_DEPENDENCIES project-wide.

* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/ThirdParty/ANGLE/ANGLE.xcodeproj/project.pbxproj:
* Source/ThirdParty/gtest/xcode/gtest.xcodeproj/project.pbxproj:
* Source/ThirdParty/libwebrtc/libwebrtc.xcodeproj/project.pbxproj:
  Previously libwebrtc was only found implicitly when building platforms
  that use webrtc, i.e. non-watchOS platforms. When building with scheme
  scopeing, its "All" aggregate needs to be part of the scheme. But we
  still don't want it to build on watchOS. Add platform filters to every
  member of the "All" aggregate target to express this.
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebGPU/WebGPU.xcodeproj/project.pbxproj:
* Source/WebInspectorUI/WebInspectorUI.xcodeproj/project.pbxproj:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Source/WebKitLegacy/WebKitLegacy.xcodeproj/project.pbxproj:
* Source/bmalloc/bmalloc.xcodeproj/project.pbxproj:
* Tools/DumpRenderTree/DumpRenderTree.xcodeproj/project.pbxproj:
* Tools/MobileMiniBrowser/MobileMiniBrowser.xcodeproj/project.pbxproj:

* Tools/TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj: Also
  remove the embedded references to gtest.xcodeproj. By changing "All
  WebKit Tools" to use scheme scope, we can also enable implicit
  dependency resolution, without causing the scheme to pull in all of
  WebKit. This means we don't have to declare nested target dependencies
  in TestWebKitAPI, just to make it always build after gtest.

* Tools/WebKitTestRunner/WebKitTestRunner.xcodeproj/project.pbxproj:
* WebKit.xcworkspace/xcshareddata/xcschemes/All WebKit Tools.xcscheme:
  See above note on enabling implicit dependencies.