[Modules] Move <AppServerSupport/OSLaunchdJob.h> of WTF headers https://bugs.webkit.org/show_bug.cgi?id=293272
Reviewed by NOBODY (OOPS!).

We use this header for webpushd, but expose it through WTF. When WTF is
imported as a module (i.e. when compiling swift), the entire
AppServerSupport module is imported, and it fails the build with an
error indicating that the framework is only supposed to be used by XPC
servers.

Fix by moving the OSLaunchdJob.h import out of XPCSPI.h and into the two
places where it's actually used. This will become a problem again if
webpushd starts building with modules enabled.

* Source/WTF/wtf/spi/darwin/XPCSPI.h:
* Source/WebKit/webpushd/webpushtool/WebPushToolMain.mm:
* Tools/TestWebKitAPI/cocoa/DaemonTestUtilities.mm: