Remove CredentialUpdaterShim interface from open source builds that will never be used https://bugs.webkit.org/show_bug.cgi?id=295500 rdar://155192244
CredentialUpdaterShim is implemented via WebKitAdditions, but its
interface is defined in open-source WebKit code. In the public build, we
load the class dynamically via soft-linking to WebKitSwift, and then
call methods on the class object.

This should always be a message send to nil, and these message sends
without a corresponding implementation confuses SPI checking. Fix by
compiling out the calls entirely until this code is upstreamed.

* Source/WebKit/UIProcess/WebAuthentication/Cocoa/WebAuthenticatorCoordinatorProxy.mm: