[webkitapipy] Allowed selector names are ambiguous between classes, leading to false positives https://bugs.webkit.org/show_bug.cgi?id=296133
Reviewed by NOBODY (OOPS!).

It's possible for a selector to be implemented multiple times in the
SDK, where some implementations are SPI methods and others are API. In
this situation, audit-spi would prevent us from listing that selector as
an allowed declaration; it knows of the API method and emits an
UnnecessaryAllowedName diagnostic.

But if we are indeed binding to an SPI method, we want to be able to
track it in the allowlist, even if the selector's access is ambiguous.
To support this, teach the allowlist format to accept an optional
class-bound selector name. Instead of a string selector:

    "initWithQueue:"

we can write:

    { name = "initWithQueue:", class = "CoreTelephonyClient" }

to allow -[CoreTelephonyClient initWithQueue:] specifically.

To support this, have APIReport (binary analysis) and SDKDB (SDK
analysis) ingestions track the class that each known selector is
implemented on. Add a clause to the audit() query to only match an
allowlist entry with a class binding if the `class_name`s match.

* Source/WebKit/Configurations/AllowedSPI-legacy.toml: Clarify some
  existing allowed selectors to be bound to specific SPI methods. These
  were discovered during work to support SPI checking in iOS 26 -- the
  selectors are all SPI methods, but there are unrelated API methods
  with the same name.

* Tools/Scripts/libraries/webkitapipy/webkitapipy/allow.py:
(AllowedSPI):
(AllowedSPI.Selector):
(AllowList):
(AllowList.from_dict): As a drive-by enhancement, fail when an allowlist entry has
  an unrecognized key. This helps catch typos instead of silently
  ignoring them.
(AllowList.from_file): As a drive-by enhancement, report the file being
  parsed when handling a decode error.
* Tools/Scripts/libraries/webkitapipy/webkitapipy/allow_unittest.py:
* Tools/Scripts/libraries/webkitapipy/webkitapipy/macho.py:
(APIReport):
(APIReport.Selector):
(APIReport._populate_from_dyld_info):
* Tools/Scripts/libraries/webkitapipy/webkitapipy/sdkdb.py:
(SDKDB._initialize_db):
(SDKDB.add_partial_sdkdb):
(SDKDB):
(SDKDB._add_partial_sdkdb):
(SDKDB.InsertionKind.statement):
(SDKDB._add_api_report):
(SDKDB._add_allowlist):
(SDKDB.add_for_auditing):
(SDKDB.audit):
(SDKDB._add_symbol):
(SDKDB._add_objc_class):
(SDKDB._add_objc_selector):
* Tools/Scripts/libraries/webkitapipy/webkitapipy/sdkdb_unittest.py: In
  addition to some new test cases, refactor the allowlist fixture logic
  to make it easier to test with different allowlists. Add cases
  which ingest SDKDBs.