[TAPI] libpas headers emit warnings on arm64_32 https://bugs.webkit.org/show_bug.cgi?id=301448 rdar://163366617
Reviewed by NOBODY (OOPS!).

libpas is disabled on this architecture, but we still pass its headers
to TAPI for InstallAPI verification, because they *are* used when
building arm64e. The result is that TAPI emits warnings when parsing for
arm64_32, like:

    In file included from WebKitBuild/Release-watchos/usr/local/include/bmalloc/pas_segregated_size_directory.h:34:
    WebKitBuild/Release-watchos/usr/local/include/bmalloc/pas_compact_tagged_page_granule_use_count_ptr.h:34:1: warning: shift count >= width of type [-Wshift-count-overflow]
       34 | PAS_DEFINE_COMPACT_TAGGED_PTR(pas_page_granule_use_count*,
          | ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
       35 |                               pas_compact_tagged_page_granule_use_count_ptr);
          |                               ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

It is not possible to give TAPI different headers for different
architectures. Instead, work around by excluding libpas headers from
TAPI on *all* builds of watchOS, and then redeclaring the actual API
surface in ExtraSymbolsForTAPI.

This trades accuracy of the InstallAPI verification on arm64e watchOS
for a cleanup on arm64_32.

* Source/JavaScriptCore/API/ExtraSymbolsForTAPI.h:
* Source/JavaScriptCore/Configurations/JavaScriptCore.xcconfig: