[Xcode] Build settings passed to `make` should forward to xcodebuild https://bugs.webkit.org/show_bug.cgi?id=303566 rdar://problem/165857240
We have handwritten Makefile logic to pass specific common build
settings like SDKROOT and ARCHS through to xcodebuild, but passing
arbitrary build settings requires using "ARGS". Use a bit of Make
metaprogramming to analyze the $(.VARIABLES) which originate from
command-line overrides. Any variables that are not part of a known set
that control Makefile logic should be passed through to xcodebuild.

For example, controlling the macOS deployment target would previously
require an invocation like:

    make debug ARGS="MACOSX_DEPLOYMENT_TARGET=15.4"  # old way

Now it can be passed more fluently like:

    make debug MACOSX_DEPLOYMENT_TARGET=15.4         # new way

Add a diagnostic which recognizes build-setting-like words in $(ARGS)
and emit a warning to  encourage the new style.

* Makefile.shared: